requirements.txt
flask
python-dotenv
openai
aiohttp
requests
gunicorn

flask
python-dotenv
requests
gunicorn
aiohttp
pymongo
dnspython
pandas
numpy
scipy
reportlab